108 create story board and tests (#117)
* Create story board and tests#108

Updated css formatting for larger screens that the footer is sticky.

* Create story board and tests#108

Updated css formatting for larger screens so that the footer is sticky.
Refactored email verification to be under signup, adjusted routes.
Added stories and test cases.

* Create story board and tests#108

Updated css formatting for larger screens so that the footer is sticky.
Refactored email verification to be under signup, adjusted routes.
Added stories and test cases.