111 create client side logic for email code verification (#119)
* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.

* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.

* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.

* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.

* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.

* Create client side logic for email code verification#111

Added client side logic for checking that the code is 6 digits.
Added UserContext to store the state of the user and added private routing for pages.