Add API Calls and logic for email#118 (#134)
* Add API Calls and logic for email#118

Added service to perform http requests to the backend
Added config to use env variables
Added axios package to perform http requests
Added logic for Email collection and code verification when calling backend
Added notice when to top when a new code is requested
Updated the button component to be disabled when the form is awaiting the response from backend
Added trxnId to the User context for code validation

* Add API Calls and logic for email#118

Added logic to populate email and code if user uses the back/forward button

* Add API Calls and logic for email#118

Updating error messages for when the server does not respond. Translation is needed.

* Add API Calls and logic for email#118

Updated validation on email collection response to the Response model from the backend.

* Add API Calls and logic for email#118

Updated error message when server is not responding.

Updated message for code resent in the notice component to use the page content for translation.

* Add API Calls and logic for email#118

Removed failing test case
Updated auth services and endpoints to work with backend
Updated backend checks for the responses to match what is currently returned.
Updated to remove the request new code if it is awaiting a response from the backend.

* Add API Calls and logic for email#118

Updating FR for password page merge.

* Add API Calls and logic for email#118

Updating FR for password page merge.

* Add API Calls and logic for email#118

Updating error message for when server is not available.

* Add API Calls and logic for email#118

Adding in log for config url testing.