Updated review changes
Need to update the link on FirstTimeGc Component to point to your page and not sign up.
The page/component names I think it is better if it is changed from profile to privacy (just privacy everywhere) since it is a privacy page and moved to a Privacy folder under Components in case it is used again.
The story should use Page to display the page and the Component under view can be deleted since we are using Page to display the component.
Might also need to update the private routing to only show the sign up page if they have already been to the privacy page.
The email link can you update to have mailto:recipient@example.com as the href?
Below i think the text is missing a space