290 sign in  create account landing/privacy notice (#299)
* Sign up - Email Collection - Accessibility Tests#233

Added Addon for accessibility in story book
Tested Accessibility with lighthouse and Color contrast extensions
Tested Color blindness and blur screen
Manually tested the Keyboard navigations.

* Create Profile component and test case

Added create profile privacy component.
Added to storybook
Added test case.
updated the En and Fr json files.

* updated what is gc sign in code

* Added test cases using test suite

Updated the test case using test suite
updated en and fr json files accordingly to fix the test case errors

* Added Missing line for test case

* removed older version test case- verification page

* 1,3,5,6 items are changed

1. Need to update the link on FirstTimeGc Component to point to your page and not sign up.
2. The page/component names I think it is better if it is changed from profile to privacy (just privacy everywhere) since it is a privacy page and moved to a Privacy folder under Components in case it is used again.
3. The story should use Page to display the page and the Component under view can be deleted since we are using Page to display the component.
4. Might also need to update the private routing to only show the sign up page if they have already been to the privacy page.
5. The email link can you update to have mailto:recipient@example.com as the href?
6. Below i think the text is missing a space

* Updated review changes

Need to update the link on FirstTimeGc Component to point to your page and not sign up.
The page/component names I think it is better if it is changed from profile to privacy (just privacy everywhere) since it is a privacy page and moved to a Privacy folder under Components in case it is used again.
The story should use Page to display the page and the Component under view can be deleted since we are using Page to display the component.
Might also need to update the private routing to only show the sign up page if they have already been to the privacy page.
The email link can you update to have mailto:recipient@example.com as the href?
Below i think the text is missing a space

* submit button

* Added Storybook test case

* review changes

Fixed the undefined error from merge commit
Updated the links to italic according to figma

* Fixed review comments

* added missing route

* Updated isEmailValid Function